Move headers in a subfolder to help integration in another project. Add the possibility to send a NOP command (useful for debug purpose or to workaround bad NIC behavior). Fix: if link layer believed  that a frame was missing and was wrong, datagraéms can be associated to wrong handler.

+++ include/kickcat/protocol.h
@@ -92,7 +92,7 @@ namespace kickcat
         }
     }
 
-    char const* ALStatus_to_string(uint16_t code);
+    char const* ALStatus_to_string(int32_t code);
 
     //TODO need unit test on bitfield to check position !
